split up fetch-page-data.ts
